restructured the OO hierarchy so that the methods/classes to grab inputs and send outputs are completely modular. This will ready us to port the wrapper on whatever BOINC/Lisp interface comes out of the reworked architecture.
I've got the methods/classes for the HPCs hardcoded in the wrapper. If 
we want to substitute these for MM methods/classes, then we just specify
a different 'sessionBuilder= ' line in the config file, and then write 
the function that that line points to. I'm currently doing this by
loading a separate 'mm-class.lisp' file, and then putting all the MM
methods/classes/functions in there.

For an HPC example, check out the 'build-hpc-session function close to the 
bottom of the wrapper. Notice how you can expand the slot definitions
when extending the OO hierarchy and still have the 'build-session' macro
write code that initializes those expanded slots. Lisp just rocks :)


git-svn-id: file:///home/cstanley/src/svnTest@175 3d771ffb-ea3b-0410-94bf-270a47e0303e