[1] added the ability to refer to IVs on lines in the config file that generate values for DVs. Figured this might be useful if we ever start doing regression testing.
[2] added the ability to print information collected at the session
level That is, there's a new collector object that gathers all the
evaluated information for the collapsed results for each line in the
work file. You can write a 'print-collector' method at this session
level, or at the level of each line in the work file (as before), and
then build a session object that will call those methods. I needed
this when building a regression test for the wrapper using results
from the DSF test, so I figured it may be a common pattern that people
would like to use.

[3] cleaned up the 'build-session' macro. You now pass three
arguments: [1] a constructor for the input class, [2] a constructor
for the collector class at the session level, and [3] a constructor
for the collector class at the level of each line in the work file.
Any of these can be omitted, and the macro will default to printing
nothing (for [2] or [3]) or running nothing (for [1]).

[4] Added a 'print-collector' method that you can define to determine
how the wrapper will output the last N lines if the model fails. For
the hpc stuff, I'm (currently) having the wrapper just print those
last 200 lines to standard output if there's an error. This
print-collector method is configured, just like the others, by adding
it as a parameter to the 'generate-session macro.

[5] Cleaned up the 'generate-session macro; should be a bit more
readable now, although the macro is designed so that I don't ever have
to really program at this level again. The wrapper is configured by
passing the objects that you want to use to the 'generate-session
macro.

[6] General code cleanup


git-svn-id: file:///home/cstanley/src/svnTest@176 3d771ffb-ea3b-0410-94bf-270a47e0303e