[1] added a 'verbose macro. If you wrap a line in the config file with this command, the string expansion of that line will be printed to the terminal before that expansion is evaluated. Could be useful for debugging.
For example, if you have this config file:

obsrt1=1.1
obsrt2=2.2
obsrt3=3.3
correlRT=(correl (list [modRT1:3]) (list [obsRT1:3]))
dv=correlRT
collapseQuota= 2

And you want to make sure that 2 runs of the model is actually
happening, you could change the config file to this:

obsrt1=1.1
obsrt2=2.2
obsrt3=3.3
correlRT=(verbose (correl (list [modRT1:3]) (list [obsRT1:3])))
dv=correlRT
collapseQuota= 3

This config file will do exactly the same thing as the config file
above, but this additional information will be printed to the terminal
as a side effect:

(correl (list (collapse (list 1.1 1.11 1.1) 'mean) (collapse (list 2.2
2.22 2.2) 'mean) (collapse (list 3.3 3.33 3.3) 'mean)) (list 1.1 2.2
3.3))

When that line is evaluated by the lisp reader, the scalar value is
bound to the variable correlRT. So, the string expansions like this
printed to terminal show what you actually wrote in the config file
for a particular variable's value (correlRT for this example).


[2] before starting the run, letf prints to the error stream all of
the uncommented lines in the config file that it didn't read. This is
useful for catching typos. For example, if you accidentally use the
word 'sessionBiulder=' instead of 'sessionBuilder=' in the config
file, letf will say that it didn't read the 'sessionBiulder=' line...


git-svn-id: file:///home/cstanley/src/svnTest@180 3d771ffb-ea3b-0410-94bf-270a47e0303e