I've changed the format for the entry function that letf requires. Instead of a hash table, if you have IVs IV1 and IV2, you write your entry function like:
(defun run-model (&key (IV1 nil) (IV2 nil))
  ())

The 'nils are just the defaults that will be bound to the variables if
no value is passed when calling the function.

To call the entry function, you'd just type

(run-model :iv1 5 :iv2 4)

I'm keeping code supporting the hash table version in there, just in
case you want backward compatibility. If you want the old version,
where the entry function needs to accept a hash table as input, just
add the 'entryFnType=hash' line in the config file (the default is
'entryFnType=keys').

Let me know if you run into any issues using this. Or, if you have any
other suggestions for the format. We're trying to make the format of
the entry function as simple as possible for the modelers, without
sacrificing too much functionality... Should have just listened to you
from the beginning...

[2] there was a bug in the 'restructure function, that reattaches
lines when the first line has a '\' at the end of it.

I found that when the banner (info at the bottom of the program) for
textWrangler about the config file says "windows (clrf)", a 'return'
typed in textwrangler actually adds both a #\linefeed and #\return
character, but when the banner says "unix (lf)", it only adds a
#\linefeed.

I was originally just skipping over the '\' and the character after
the '\', but adding multiple return statements after the '\' broke the
code...

So, the 'restructure function now has a while loop after finding a '\'
at the end of a line, and steps over all of the (#\linefeed #\return
#\newline) characters until finding something else, or the end of the
entire string. Then it starts appending the remaining characters of
the string to the line with the '\' at the end of it, until finding
another line terminator.

letf is now ready for server-side validation of the config file (FB case
816). mm.lisp has the few extra methods to  have letf validate the
config file without having to run the model, or read in a work file.


git-svn-id: file:///home/cstanley/src/svnTest@184 3d771ffb-ea3b-0410-94bf-270a47e0303e