I started digging into the code optimization section of Hoyte's book, and I took a day to apply some of the tricks to speed up letf. It should be around an order of magnitude faster now, both loading the config file and during the model runs.
To get the speed up, I rewrote a few of the algorithms, but mostly
just used type-specific function calls (instead of type-agnotist
calls). Easy fixes like using 'string-equal instead of 'equalp... I
also wrote a 'concatenate-fast macro that has the same functionality
as 'concatenate, but leverages the low-level 'with-output-to-string
macro to get ~5x speedup.


git-svn-id: file:///home/cstanley/src/svnTest@190 3d771ffb-ea3b-0410-94bf-270a47e0303e