[1] Made improvements to the interface between letf and the model, when the model is launched as a separate process.
When the model is launched as a separate process, it sends its outputs
to stdout, and grabs its inputs from stdin (not from a command-line
argument)

For example, if you have runsPerProcess=3 and 4 IVs in the config
file, when the model is launched, it's stdin might look like:

1       2       3       5
1       2       4       5
1       2       5       5

How each row in stdin looks is configurable. If you want to change the
format of a row (say make it look like this):

IV1=1   IV2=2   IV3=3   IV4=5

You add a line in the config file:

IVStringFn=#'new-function

Where 'new-function is your custom method that takes a run object as
input, and returns a string (which is a row in standard input) as
output. Check 'defaultIVStringFn as an example. This method is what
the code will default to if you don't specify an IVStringFn in the
config file.


git-svn-id: file:///home/cstanley/src/svnTest@193 3d771ffb-ea3b-0410-94bf-270a47e0303e